feat: refactor CreateTestCase to use resolution-helper and improve field resolution logic
